Removed unnecessary nesting of package names (yoyo.migrate.* -> yoyo.*)
The extra level of nesting added nothing, so I'm taking the opportunity to
remove it and simplify the source code layout.

This will break things for anyone using the package API (ie importing functions
from yoyo.migrate), but it should be a simple search and replace to fix this.
In any case I think most users will be using the cli script, which remains
unaffected.